pyproject.toml
[tool.poetry]
name = "pdf-quiz-maker-backend"
version = "0.1.0"
description = ""
authors = ["KeishiIrisa <keishi.irisa@gmail.com>"]
readme = "README.md"
packages = [{include = "pdf_quiz_maker_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
llama-index = ">=0.12.0,<0.13.0"
openai = "^1.53.0"
python-dotenv = "^1.0.1"
pymongo = {version = "3.11", extras = ["srv"]}
astrapy = ">=1.5,<2.0"
llama-index-vector-stores-astra-db = "^0.4.0"
certifi = "^2024.8.30"
markdown = "^3.7"
llama-index-readers-file = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# 1行の最大文字数
line-length = 88

[tool.ruff.lint]
# チェックするエラーの種類
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
# 除外するエラーの種類
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "B904", # raise ... from ... になっていない場合のエラーを無視
]

# ファイルごとのエラー除外
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

# isort の設定
[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.lint.pyupgrade]
# Python3.8互換のための設定
keep-runtime-typing = true
